# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Blockchain Gateway Backend API
  version: 3.0.0
  description: |
    Backend API for chaincode lifecycle orchestration with RBAC.
    
    This API provides comprehensive management of:
    - User authentication and authorization with JWT tokens
    - Role-based access control (ADMIN, ORG_ADMIN, USER, VIEWER)
    - Chaincode lifecycle management (upload, validate, approve, deploy)
    - Deployment orchestration with workflow management
    - Certificate synchronization with Fabric CA
    - Audit logging for all operations
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Roles and Permissions
    
    - **ADMIN**: Full system access, can manage users, approve chaincodes, configure system
    - **ORG_ADMIN**: Organization administrator, can upload/deploy chaincodes, view users
    - **USER**: Regular user, can invoke/query chaincodes, manage assets
    - **VIEWER**: Read-only access, can query chaincodes and view assets
    
  contact:
    name: Blockchain Gateway Team
    email: support@blockchain.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.blockchain.com/api/v1
    description: Production server
...
paths:
...
  /users:
    post:
      tags:
        - User Management
      summary: Create new user
      description: Create a new user (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - User Management
      summary: Get users list
      description: Get list of users with optional filters (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: skip
          in: query
          description: Number of records to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of records to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: role
          in: query
          description: Filter by role
          required: false
          schema:
            type: string
            enum: [ADMIN, ORG_ADMIN, USER, VIEWER]
        - name: status
          in: query
          description: Filter by status
          required: false
          schema:
            type: string
            enum: [active, inactive, suspended]
        - name: organization
          in: query
          description: Filter by organization
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{user_id}:
    get:
      tags:
        - User Management
      summary: Get user by ID
      description: Get user information by ID (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - User Management
      summary: Update user
      description: Update user information (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Chaincode Management endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        detail:
          type: string
          example: "Detailed error information"

    # Authentication schemas
...
    UserCreateRequest:
      type: object
      required:
        - username
        - email
        - password
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        password:
          type: string
          minLength: 8
          example: "securepassword123"
        role:
          type: string
          enum: [ADMIN, ORG_ADMIN, USER, VIEWER]
          example: "USER"
        msp_id:
          type: string
          example: "Org1MSP"
        organization:
          type: string
          example: "Example Corp"
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [ADMIN, ORG_ADMIN, USER, VIEWER]
          example: "USER"
        msp_id:
          type: string
          example: "Org1MSP"
        organization:
          type: string
          example: "Example Corp"
        status:
          type: string
          enum: [active, inactive, suspended]
          example: "active"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          example: "john@example.com"
        role:
          type: string
          enum: [ADMIN, ORG_ADMIN, USER, VIEWER]
          example: "USER"
        msp_id:
          type: string
          example: "Org1MSP"
        organization:
          type: string
          example: "Example Corp"
        status:
          type: string
          enum: [active, inactive, suspended]
          example: "active"
        is_active:
          type: boolean
          example: true
        is_verified:
          type: boolean
          example: false
        last_login:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00Z"
        created_at:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00Z"
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        total:
          type: integer
          example: 25
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 100

    # Chaincode Management schemas
...